+++ packages/connect-wallet-modal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reef-knot/connect-wallet-modal
 
+## 5.5.1
+
+### Patch Changes
+
+- Fix the issue with paused loading animation for ConnectButtonLoader
+
 ## 5.5.0
 
 ### Minor Changes