+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,29 @@
 # reef-knot
 
+## 5.7.3
+
+### Patch Changes
+
+- Fix the issue with paused loading animation for ConnectButtonLoader
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@5.5.1
+
+## 5.7.2
+
+### Patch Changes
+
+- Fix: onAccountsChanged, onChainChanged methods for ledgerLiveConnector
+- Updated dependencies
+  - @reef-knot/ledger-connector@4.1.2
+
+## 5.7.1
+
+### Patch Changes
+
+- Fix: wrap IFrameEthereumProvider in LedgerIFrameProvider with the "request" method, which is required for useWalletClient wagmi hook
+- Updated dependencies
+  - @reef-knot/ledger-connector@4.1.1
+
 ## 5.7.0
 
 ### Minor Changes